+++ RELEASE_NOTES.md
@@ -1,6 +1,7 @@
 ### New in 6.1.0 (Unreleased)
 
 * Make IMonitorRegistry generic. This fixes a bug where behaviour is different depending on if the registry is stored as IMonitorRegistry or OkanshiMonitorRegistry
+* Okanshi.Owin now has an option for which timer to use when monitoring.
 * `SlaTimer` has been added. The timer makes it easy to track SLA-breaches.
 * `MinMaxAvgGauge` has been added. It allows you to detect extreme values that would otherwise disappear in an average calculation.
 * `OkanshiMonitor.DefaultStep` has been deleted since it was forgotten in an earlier clean up process.